backend/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.11.0
kafka-python==2.0.2
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.8
redis==4.2.2
sqlparse==0.4.2
wrapt==1.14.0
